app/requirements.txt
python-socketio[client]
websocket-client
stable-baselines3[extra]
gymnasium[box2d]